docsearch-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
faiss-cpu
google-generativeai
PyPDF2
python-docx